Add support for disabling text/button widgets
A disabled button cannot be clicked, selected (it wouldn't react to
hovering over it with the mouse) and has a distrinct text color.

This is useful for keeping a button visable (to keep other widgets
aligned) without letting the player click on it.  This will be needed
in a future commit.

Signed-off-by: Niels Thykier <niels@thykier.net>